Removed UtcTypeTrait
