+++ DBAL/Types/UtcDateType.php
@@ -25,8 +25,6 @@ use Doctrine\DBAL\Types\DateType;
  */
 class UtcDateType extends DateType
 {
-    use UtcTypeTrait;
-
     /**
      * {@inheritdoc}
      */
@@ -57,5 +55,25 @@ class UtcDateType extends DateType
     public function requiresSQLCommentHint(AbstractPlatform $platform): bool
     {
         return true;
+    }
+
+    /**
+     * @throws ConversionException
+     */
+    private function convertToDateTime(?string $dateString, string $dateFormat): ?DateTime
+    {
+        if (null === $dateString) {
+            return null;
+        }
+
+        $converted = DateTime::createFromFormat($dateFormat, $dateString, new DateTimeZone('UTC'));
+
+        if (!$converted) {
+            throw ConversionException::conversionFailed($dateString, $this->getName());
+        }
+
+        $errors = $converted::getLastErrors();
+
+        return $errors['warning_count'] > 0 && (int) $converted->format('Y') < 0 ? null : $converted;
     }
 }
